+++ CHANGES.rst
@@ -9,7 +9,7 @@ Changelog
 - Fixed the way JS function `submitFormHelperOnsuccessDefault` manages
   returned result when it is a file, now we have a correct `filename`.
   [gbastien]
-- Added transmogrifier Expression and Condition classes.
+- Added transmogrifier Expression and Condition classes to log expression compilation or interpretation errors.
   [sgeulette]
 
 1.0.0rc1 (2024-02-08)